+++ ASolver.hh
@@ -2,6 +2,7 @@
 
 #include <string>
 #include <queue>
+#include <cstdint>
 
 #include "Solver.hh"
 
@@ -12,10 +13,10 @@ namespace TaquinSolve
         public:
             ASolver();
 
-            std::queue<Moves> solve(std::vector<size_t> board, size_t board_size);
+            std::queue<Moves> solve(std::vector<uint8_t> board, uint8_t board_size);
             std::vector< std::shared_ptr<Board> > perform_moves(Board *board, std::vector<Moves> moves);
 
         protected:
-            std::shared_ptr<Board> get_cheapest_board(std::map<size_t, std::shared_ptr<Board> > *open_set);
+            std::shared_ptr<Board> get_cheapest_board(std::map<uint64_t, std::shared_ptr<Board> > *open_set);
     };
 }
