+++ Solver.hh
@@ -4,6 +4,7 @@
 #include <vector>
 #include <map>
 #include <string>
+#include <cstdint>
 
 #include "Board.hh"
 
@@ -14,11 +15,11 @@ namespace TaquinSolve
         public:
             Solver();
 
-            virtual std::queue<Moves> solve(std::vector<size_t> board, size_t board_size) = 0;
+            virtual std::queue<Moves> solve(std::vector<uint8_t> board, uint8_t board_size) = 0;
             virtual std::vector< std::shared_ptr<Board> > perform_moves(Board *board, std::vector<Moves> moves) = 0;
 
         protected:
-            std::shared_ptr< std::map<size_t, int> > pattern_database = NULL;
+            std::shared_ptr< std::map<uint64_t, uint8_t> > pattern_database = NULL;
 
         private:
             void load_database(std::string path);
